+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * Removed `Claims` class (deprecated since 1.0)
 * Removed `Entity::getClaims` (deprecated since 1.0)
 * Removed `HashableObjectStorage`
+* Removed `getLabels`, `getDescriptions`, `setLabels` and `setDescriptions` from `Entity`
 
 ## Version 4.4.0 (2016-01-20)